--- .vscode/settings.json
@@ -1,24 +1,0 @@
-{
-	"files.associations": {
-		"parser.h": "c",
-		"libft.h": "c",
-		"history.h": "c",
-		"ioctl.h": "c",
-		"readline.h": "c",
-		"stdlib.h": "c",
-		"typeinfo": "c",
-		"exec.h": "c",
-		"__locale": "c",
-		"atomic": "c",
-		"cstdlib": "c",
-		"locale": "c",
-		"new": "c",
-		"ostream": "c",
-		"algorithm": "c",
-		"complex": "c",
-		"array": "c",
-		"string": "c",
-		"string_view": "c",
-		"vector": "c"
-	}
-}
