+++ src/core/exec/m_builtins1.c
@@ -1,4 +1,16 @@
-#include "exec.h"
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   m_builtins1.c                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ahkiler <ahkiler@student.42.fr>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/03/23 19:36:12 by sechung           #+#    #+#             */
+/*   Updated: 2023/03/25 13:58:24 by ahkiler          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
 
 void	m_echo(t_process *proc, t_deque *argv, int flag)
 {
@@ -29,7 +41,7 @@ t_token	*check_echo_option(t_deque *argv, int *option)
 
 	*option = 0;
 	str = argv->top->next;
-	while (str != NULL && !ft_strcmp ("-n", str->content))
+	while (str != NULL && is_option (str->content))
 	{
 		*option = 1;
 		str = str->next;
@@ -37,12 +49,28 @@ t_token	*check_echo_option(t_deque *argv, int *option)
 	return (str);
 }
 
+int	is_option(char *content)
+{
+	int	i;
+
+	i = 0;
+	if (content == NULL)
+		return (0);
+	if (content[i++] != '-')
+		return (0);
+	while (content[i])
+	{
+		if (content[i++] != 'n')
+			return (0);
+	}
+	return (1);
+}
+
 void	m_env(t_process *proc, int flag)
 {
 	t_token	*env;
 	char	*content;
 	int		env_flag;
-	int		i;
 
 	env = proc->envp->top;
 	env_flag = 0;
