+++ src/core/init/init_data.c
@@ -1,25 +1,21 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_strcmp.c                                        :+:      :+:    :+:   */
+/*   init_data.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: sechung <sechung@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/03/23 19:57:43 by sechung           #+#    #+#             */
-/*   Updated: 2023/03/23 19:58:06 by sechung          ###   ########.fr       */
+/*   Created: 2023/03/23 19:48:20 by sechung           #+#    #+#             */
+/*   Updated: 2023/03/23 20:18:06 by sechung          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "minishell.h"
 
-int	ft_strcmp(const char *s1, const char *s2)
+void	init_data(t_data *data)
 {
-	size_t	i;
-
-	i = 0;
-	if (s1 == s2)
-		return (0);
-	while (s1[i] && s1[i] == s2[i])
-		i++;
-	return ((unsigned char) s1[i] - (unsigned char) s2[i]);
+	data->envs = 0;
+	data->line = 0;
+	data->tokens = 0;
+	data->tree = 0;
 }
