--- src/core/lexer/init_data.c
@@ -1,9 +1,0 @@
-#include "parser.h"
-
-void	init_data(t_data *data)
-{
-	data->envs = 0;
-	data->line = 0;
-	data->tokens = 0;
-	data->tree = 0;
-}
