+++ src/core/signal/signal.c
@@ -1,6 +1,16 @@
-#include "parser.h"
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   signal.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ahkiler <ahkiler@student.42.fr>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/03/23 19:54:07 by sechung           #+#    #+#             */
+/*   Updated: 2023/03/25 13:44:28 by ahkiler          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
-extern int	g_exit_status;
+#include "minishell.h"
 
 void	new_prompt_handler(int signo)
 {
@@ -10,19 +20,16 @@ void	new_prompt_handler(int signo)
 		write (1, "\n", 1);
 		rl_on_new_line();
 		rl_redisplay();
+		g_exit_status = 1;
 	}
 }
 
 void	child_signal_handler(int signo)
 {
-	// write(1, "cc\n", 3);
-	// if (signo == SIGINT)
-	// ft_putendl_fd("", STDERR_FILENO);
-	// if (signo == SIGQUIT)
-	// 	ft_putendl_fd("Quit: 3", STDOUT_FILENO);
+	(void) signo;
+	ft_putendl_fd("", STDERR_FILENO);
 	exit(1);
 }
-
 
 void	default_signal(void)
 {
@@ -30,7 +37,7 @@ void	default_signal(void)
 	signal(SIGQUIT, SIG_IGN);
 }
 
-void	set_herdoc_signal(pid_t pid)
+void	set_herdoc_signal(void)
 {
 	signal(SIGINT, child_signal_handler);
 }
@@ -39,13 +46,8 @@ void	set_signal(pid_t pid)
 {
 	if (pid == 0)
 	{
-		// write(1, "hh\n", 3);
 		signal(SIGINT, child_signal_handler);
-		// if (heredoc)
-		// 	signal(SIGQUIT, SIG_IGN);
-		// else
 		signal(SIGQUIT, child_signal_handler);
-		// write(1, "\n", 1);
 	}
 	else
 		signal(SIGINT, SIG_IGN);
