+++ src/core/token/token.c
@@ -1,25 +1,35 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_strcmp.c                                        :+:      :+:    :+:   */
+/*   token.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: sechung <sechung@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/03/23 19:57:43 by sechung           #+#    #+#             */
-/*   Updated: 2023/03/23 19:58:06 by sechung          ###   ########.fr       */
+/*   Created: 2023/03/23 19:54:52 by sechung           #+#    #+#             */
+/*   Updated: 2023/03/23 20:19:12 by sechung          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "minishell.h"
 
-int	ft_strcmp(const char *s1, const char *s2)
+t_token	*get_token(void)
 {
-	size_t	i;
+	t_token	*token;
 
-	i = 0;
-	if (s1 == s2)
-		return (0);
-	while (s1[i] && s1[i] == s2[i])
-		i++;
-	return ((unsigned char) s1[i] - (unsigned char) s2[i]);
+	token = 0;
+	while (!token)
+		token = ft_calloc(1, sizeof(t_token));
+	while (!(token->content))
+		token->content = ft_calloc(1, sizeof(char));
+	return (token);
+}
+
+t_token	*get_env_token(void)
+{
+	t_token	*token;
+
+	token = 0;
+	while (!token)
+		token = ft_calloc(1, sizeof(t_token));
+	return (token);
 }
