+++ src/core/utils/utils.c
@@ -1,25 +1,36 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_strcmp.c                                        :+:      :+:    :+:   */
+/*   utils.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: sechung <sechung@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/03/23 19:57:43 by sechung           #+#    #+#             */
-/*   Updated: 2023/03/23 19:58:06 by sechung          ###   ########.fr       */
+/*   Created: 2023/03/23 19:55:18 by sechung           #+#    #+#             */
+/*   Updated: 2023/03/23 20:19:15 by sechung          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "minishell.h"
 
-int	ft_strcmp(const char *s1, const char *s2)
+char	*join_line(char *dest, char *s)
 {
-	size_t	i;
+	char	*new_line;
+	char	*merge_line;
 
-	i = 0;
-	if (s1 == s2)
-		return (0);
-	while (s1[i] && s1[i] == s2[i])
-		i++;
-	return ((unsigned char) s1[i] - (unsigned char) s2[i]);
+	new_line = 0;
+	merge_line = 0;
+	while (!merge_line)
+		merge_line = ft_strdup(s);
+	while (!new_line)
+		new_line = ft_strjoin(dest, merge_line);
+	free(dest);
+	free(merge_line);
+	return (new_line);
+}
+
+int	white_space(char c)
+{
+	if ((c >= 9 && c <= 13) || c == 32)
+		return (1);
+	return (0);
 }
